Clarify "data does not exist" vs empty data
The trips and status changes endpoints are expected to return
HTTP 404 responses when data "does not exist". I'm trying to
clarify that that means the future and hours when the provider
was not in operation in the municipality. For past hours during
which the provider was operating but nothing happened they should
return empty arrays of trips or status changes.